backend/requirements.txt
Flask
Flask-CORS
python-dotenv
google-generativeai
gunicorn
numpy
scipy
scikit-learn
pydantic